Clean-up, refactoring
- Log to tools
- clean-up comments adn fake data
