+++ ChitChat02/Profile/UserModel.swift
@@ -13,23 +13,5 @@ struct UserModel {
     let description: String
     var avatar: URL?
 }
-let newUser = UserModel(name: "", description: "")
 
-let fakeUser = UserModel(
-    name: "Timur Tharkahov",
-    description: "description"
-//    description: """
-//Lorem ipsum dolor sit er elit lamet, consectetaur cillium adipisicing pecu,
-//sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
-//Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo
-//consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat
-//nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt
-//mollit anim id est laborum. Nam liber te conscient to factor tum poen legum odioque civiuda.
-//Lorem ipsum dolor sit er elit lamet, consectetaur cillium adipisicing pecu,
-//sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
-//Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo
-//consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat
-//nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt
-//mollit anim id est laborum. Nam liber te conscient to factor tum poen legum odioque civiuda.
-//"""
-)
+let newUser = UserModel(name: "", description: "")
